+++ packages/expect-utils/src/utils.ts
@@ -13,6 +13,7 @@ import {
   isImmutableList,
   isImmutableOrderedKeyed,
   isImmutableOrderedSet,
+  isImmutableRecord,
   isImmutableUnorderedKeyed,
   isImmutableUnorderedSet,
 } from './jasmineUtils';
@@ -260,7 +261,8 @@ export const iterableEquality = (
   if (
     !isImmutableList(a) &&
     !isImmutableOrderedKeyed(a) &&
-    !isImmutableOrderedSet(a)
+    !isImmutableOrderedSet(a) &&
+    !isImmutableRecord(a)
   ) {
     const aEntries = Object.entries(a);
     const bEntries = Object.entries(b);